Aplikacja/backend/requirements.txt
asgiref==3.5.2
Django==4.1.3
sqlparse==0.4.3
django-environ>=0.9.0 
gunicorn==20.1.0
gevent==22.10.2
psycopg2-binary==2.9.5
certifi==2022.12.7
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
pytz==2022.6
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.13
pre-commit==2.20.0
drf_yasg==1.21.4
cffi==1.15.0
greenlet>=2.0.0
pycparser==2.21
zope.event==4.5.0
zope.interface==5.4.0
PyJWT==2.4.0
django-cors-headers==3.11.0
djangorestframework>=3.14.0